Fix MDBList directory class discovery
